agent: Move unsafe_provider to utils
