+++ agent/src/main/java/com/code_intelligence/jazzer/utils/UnsafeProvider.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.code_intelligence.jazzer.runtime;
+package com.code_intelligence.jazzer.utils;
 
 import java.lang.reflect.Field;
 import sun.misc.Unsafe;
