+++ driver/src/main/java/com/code_intelligence/jazzer/driver/BUILD.bazel
@@ -45,8 +45,8 @@ java_jni_library(
         "//agent/src/main/java/com/code_intelligence/jazzer/instrumentor",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:coverage_map",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
         "//agent/src/main/java/com/code_intelligence/jazzer/utils:manifest_utils",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:unsafe_provider",
         "//driver/src/main/java/com/code_intelligence/jazzer/autofuzz",
     ],
 )
@@ -58,7 +58,7 @@ java_jni_library(
     visibility = ["//driver:__subpackages__"],
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:unsafe_provider",
     ],
 )
