+++ driver/src/test/java/com/code_intelligence/jazzer/driver/BUILD.bazel
@@ -6,7 +6,7 @@ java_test(
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:coverage_map",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:unsafe_provider",
         "//driver/src/main/java/com/code_intelligence/jazzer/driver:fuzz_target_runner",
     ],
 )
