+++ sanitizers/src/main/java/com/code_intelligence/jazzer/sanitizers/BUILD.bazel
@@ -5,7 +5,7 @@ java_library(
     srcs = ["RegexRoadblocks.java"],
     deps = [
         "//agent:jazzer_api_compile_only",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils:unsafe_provider",
         "//sanitizers/src/main/java/com/code_intelligence/jazzer/sanitizers/utils:reflection_utils",
     ],
 )
